+++ src/main/java/be15fintomatokatchupbe/oauth/command/application/repository/InstagramPostInsightRepository.java
@@ -1,6 +1,6 @@
 package be15fintomatokatchupbe.oauth.command.application.repository;
 
-import be15fintomatokatchupbe.oauth.query.domain.InstagramPostInsight;
+import be15fintomatokatchupbe.oauth.command.application.domain.InstagramPostInsight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InstagramPostInsightRepository extends JpaRepository<InstagramPostInsight, Long> {
