+++ src/main/java/be15fintomatokatchupbe/oauth/query/service/InstagramStatsSnapshotService.java
@@ -3,8 +3,8 @@ package be15fintomatokatchupbe.oauth.query.service;
 import be15fintomatokatchupbe.influencer.command.domain.aggregate.entity.Influencer;
 import be15fintomatokatchupbe.oauth.command.application.repository.InstagramMediaSnapshotRepository;
 import be15fintomatokatchupbe.oauth.command.application.repository.InstagramStatsSnapshotRepository;
-import be15fintomatokatchupbe.oauth.query.domain.InstagramMediaSnapshot;
-import be15fintomatokatchupbe.oauth.query.domain.InstagramStatsSnapshot;
+import be15fintomatokatchupbe.oauth.command.application.domain.InstagramMediaSnapshot;
+import be15fintomatokatchupbe.oauth.command.application.domain.InstagramStatsSnapshot;
 import be15fintomatokatchupbe.oauth.query.dto.InstagramFullSnapshot;
 import be15fintomatokatchupbe.oauth.query.dto.InstagramMediaStats;
 import be15fintomatokatchupbe.oauth.query.mapper.InstagramQueryMapper;
