+++ src/main/resources/mappers/oauth/youtubeQueryMapper.xml
@@ -4,7 +4,7 @@
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="be15fintomatokatchupbe.oauth.query.mapper.YoutubeQueryMapper">
-    <select id="findYoutubeStats" resultType="be15fintomatokatchupbe.oauth.query.dto.YoutubeStatsSnapshot">
+    <select id="findYoutubeStats" resultType="be15fintomatokatchupbe.oauth.query.dto.YoutubeStats">
         SELECT
             total_videos        AS totalVideos,
             avg_views           AS avgViews,
